# api-spec.yaml
openapi: '3.0.2'
info:
  title: BizzFuzz API
  version: '0.0.1'
servers:
  - url: http://localhost:3000

paths:
  /sessions/{username}:
    get:
      summary: Validate current session for the provided username
      parameters:
        - name: username
          in: path
          required: true
          description: Username for the session
          schema:
            type: string
      responses:
        '200':
          description: OK - The session with that username exists
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: Not found - The sessions with that username does not exist
          content:
            text/plain:
              schema:
                type: string
    post:
      summary: Creates a session
      description: Validates username and password, and creates a session if a user with credentials exists
      parameters:
        - name: username
          in: path
          required: true
          description: Username to validate for session creation
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Password'
      responses:
        '201':
          description: Created - The session was created
          content:
            text/plain:
              schema:
                type: string
        '400':
          description: Bad request - Missing parts or all of requestBody
          content:
            text/plain:
              schema:
                type: string
        '403':
          description: Forbidden - Wrong credentials
          content:
            text/plain:
              schema:
                type: string
    delete:
      summary: Deletes a session
      description: Will only destroy the session if the username matches the session
      parameters:
        - name: username
          in: path
          required: true
          description: Username to delete the session
          schema:
            type: string
      responses:
        '204':
          description: No content - The session was deleted
  /users/{username}:
    get:
      summary: Retrieve all data for the user in the path parameter
      parameters:
        - name: username
          in: path
          required: true
          description: Username to get the information for
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Password'
      responses:
        '200':
          description: OK - Sends all data stored for the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserData'
        '400':
          description: Bad request - Missing requestBody
          content:
            text/plain:
              schema:
                type: string
        '403':
          description: Forbidden - Wrong credentials
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: There was an interanl error
          content:
            text/plain:
              schema:
                type: string
    post:
      summary: Create an account with the specified username and password
      description: Also creates a session for that user (so you don't need to send a login request right afterwards)
      parameters:
        - name: username
          in: path
          required: true
          description: Username to create with account
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Password'
      responses:
        '201': 
          description: Created - User successfully created 
          content:
            text/plain:
              schema:
                type: string
        '400':
          description: Bad request - missing requestBody
          content:
            text/plain:
              schema:
                type: string
        '409':
          description: Conflict - Username already taken
          content:
            text/plain:
              schema:
                type: string
    delete:
      summary: Delete an account with specified username
      description: Will fail if password doesn't match
      parameters:
        - name: username
          in: path
          required: true
          description: Username of account to be deleted
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Password'
      responses:
        '204':
          description: No Reponse - User successfully deleted
        '400':
          description: Bad request - Missing requestBody
          content:
            text/plain:
              schema:
                type: string
        '403': 
          description: Forbidden - Wrong credentials
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: There was an internal error
          content:
            text/plain:
              schema:
                type: string
    put:
      summary: Change the password of the account
      parameters:
      - name: username
        in: path
        required: true
        description: Username of account to be updated
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDataUpdateRequest'
      responses:
        '200':
          description: OK - Successfully updated
          content:
            text/plain:
              schema:
                type: string
        '400':
          description: Bad request - Missing requestBody
          content:
            text/plain:
              schema:
                type: string
        '403':
          description: Forbidden - Wrong credentials
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: There was an internal error
          content:
            text/plain:
              schema:
                type: string
  /quizlets:
    get:
      summary: Retrieve up to 9 new quizlets for given userId
      description: query field in request body is only necessary when searching. For home page, it can be left empty. Reccomended sort for home page is "MOST POPULAR"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuizletDataRequest'
      responses:
        '200':
          description: Returns up to 9 quizlet objects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizletDataResponse'
        '400':
          description: Bad request
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: There was an internal error
          content:
            text/plain:
              schema:
                type: string
  /quizzes:
    post:
      summary: Create a new quiz
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuizCreationRequest'
      responses:
        '201':
          description: Quiz successfully created
          content:
            text/plain:
              schema:
                type: string
        '400':
          description: Bad request
          content:
            text/plain:
              schema:
                type: string
        '401':
          description: Unauthorized - Not logged into any account
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: There was an internal error
          content:
            text/plain:
              schema:
                type: string
  /quizzes/{quizId}:
    get:
      summary: Get all quiz data
      parameters:
        - name: quizId
          in: path
          required: true
          description: quizId to get all quiz data
          schema:
            type: string
      responses:
        '200':
          description: Found quiz data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizDataResponse'
        '404':
          description: No quiz with that ID found
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: There was an internal error
          content:
            text/plain:
              schema:
                type: string
    delete:
      summary: delete a quiz with quizId. Requires a session with the creator of the quiz (login)
      parameters:
        - name: quizId
          in: path
          required: true
          description: quizId that corresponds with quiz to delete
          schema:
            type: string
      responses:
        '204': 
          description: No reponse - Quiz successfully deleted
        '400':
          description: Bad request
          content:
            text/plain:
              schema:
                type: string
        '401':
          description: Unauthorized - Not logged in
          content:
            text/plain:
              schema:
                type: string
        '403':
          description: Wrong credentials
          content:
            text/plain:
              schema:
                type: string


components: 
  schemas:
    Password:
      type: object
      properties:
        password:
          type: string
          example: 'password123'
    UserDataUpdateRequest:
      type: object
      properties:
        password:
          type: string
          description: the old password on the account
          example: 'oldPassword123'
        newPassword:
          type: string
          description: the new password for the account
          example: 'newPassword123'
    UserData:
      type: object
      properties:
        id:
          type: string
          description: uuid v4 prefixed with a 'u'
          example: 'u45a9d6a1-1b52-48dc-910d-dd84e1a6559c'
        username:
          type: string
          example: 'john_doe'
        dateCreated:
          type: integer
          description: unix time (seconds since epoch)
          example: '1700458466'
        completedQuizes:
          type: array
          description: a list of quizzes the user has taken along with results
          items:
            $ref: '#/components/schemas/UserDataCompletedQuiz'
    UserDataCompletedQuiz:
      type: object
      properties:
        quizId:
          type: string
          description: uuid v4 prefixed with a 'q'
          example: 'q812c47c9-8a58-49bc-8a14-4e6af6b566f0'
        dateTaken:
          type: integer
          description: unix time (seconds since epoch)
          example: '1700458954'
        result:
          $ref: '#/components/schemas/QuizDataResult'

    #Quizlet GET section
    QuizletDataRequest:
      type: object
      description: query should be empty and sort should be MOST POPULAR when loading home page, when searching, use send search term in query field, and when changing sort, use one of the enums
      properties:
        userId:
          type: string
          description: uuid v4 prefixed with a 'u'
          example: 'ubd002d37-e846-4152-9736-0f23baba5c34'
        page:
          type: integer
          description: the page number that you want (default should be 1)
          example: 2
        query:
          type: string
          description: search query, if searching
          example: harry
        sort:
          type: string
          enum:
            - MOST POPULAR
            - LEAST POPULAR
            - NEWEST
            - OLDEST
            - MOST RELEVANT
    QuizletDataResponse:
      type: array
      items:
        $ref: '#/components/schemas/QuizletData'
      minItems: 1
      maxItems: 9
    QuizletData:
      type: object
      properties:
        quizId:
          type: string
        name:
          type: string
          description: name for the quiz
          example: Harry Potter quiz
        description:
          type: string
          description: description for the quiz
          example: This is a basic quiz to find out what Harry Potter house you're in
        dateCreated:
          type: integer
          description: unix time (seconds since epoch)
          example: 1700633678
        creator:
          type: string
          description: userId of the creator of the quiz. uuid v4 prefixed with a 'u'
          example: 'ufa32357f-b118-47cc-a9dd-d0e831c2a6a9'
        taken:
          type: boolean
          description: false if the user hasn't taken the quiz before, and vice versa
          example: false
        submissions:
          type: integer
          description: the number of times the quiz has been taken
          example: 126

    #Quiz GET section
    QuizDataResponse:
      type: object
      properties:
        id:
          type: string
          description: uuid v4 prefixed with a 'q'
          example: 'q57f71583-7e73-4a22-b7fc-4b95b4a6de84'
        name:
          type: string
          description: name of the quiz
          example: 'Harry Potter Sorting House Quiz'
        description:
          type: string
          description: accompanying description for the quiz
          example: 'This quiz sorts you into the Harry Potter house that fits you best'
        dateCreated:
          type: integer
          description: unix time (seconds since epoch)
          example: '1700458954'
        creator:
          type: string
          description: uuid v4 prefixed with a 'u', aligned with the user that created the quiz
          example: '9c41e8eb-de7f-4fdc-a2b0-d88b98dc3357'
        questions:
          type: array
          items:
            $ref: '#/components/schemas/QuizDataQuestion'
        results:
          type: array
          items:
            $ref: '#/components/schemas/QuizDataResult'
    QuizDataQuestion:
      type: object
      properties:
        prompt:
          type: string
          description: the question that the quiz is asking
          example: What is your favorite color?
        answers:
          type: array
          description: a list of answers with a linked result and weight
          items:
            $ref: '#/components/schemas/QuizDataAnswer'
    QuizDataAnswer:
      type: object
      properties:
        option:
          type: string
          description: the actual choice the user sees and clicks on
          example: Red
        result:
          type: integer
          description: the result that the answer corresponds with
          example: Gryffindor
        weight:
          type: integer
          description: the weight that gets added to the result
          example: 14
    QuizDataResult:
      type: object
      properties:
        name:
          type: string
          description: the name of the result
          example: 'Hufflepuff'
        description:
          type: string
          description: a description of the result
          example: 'Congrats you''re a normie!'
    
    #Quiz POST section
    QuizCreationRequest:
      type: object
      properties:
        name: 
          type: string
          description: Name of the quiz
        description:
          type: string
          description: Description for the quiz
        creator:
          type: string
          description: the creators userId, which is a uuid v4 prefixed with a u 
          example: 'u956fbe0e-21f0-48d0-aaa8-43d213446b43'
        questions:
          type: array
          items:
            $ref: '#/components/schemas/QuizDataQuestion'
        results:
          type: array
          items:
            $ref: '#/components/schemas/QuizDataResult'